extract all values
